Updated AddItem
Added a form to AddItem.jsx for item submission.

The form handles API requests, loading state, error display, and redirects to the items page upon success.

All things are currently working as intending - adding works and redirects effectively.

Note that addItems is currently implemented as a page for learning purposes and for simplicity - good practice would be making it a component (reusability & UI standard - navigating away from a page to submit an item is generally bad practice). Will refactor as a component soon.